refactor: rename testing.py to utils.py and fix import paths
