+++ sketchlogic/model/controller.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 import sketchlogic.model.inference as inference
-import sketchlogic.model.testing as testing
+import sketchlogic.model.utils as utils
 import numpy
 import cv2
 
@@ -17,8 +17,8 @@ def run(input_image: numpy.ndarray, debug: bool = False) -> tuple[list, int]:
     results, next_id = inference.run(input_image, model_path)
 
     if debug:
-        testing.draw_results(input_image, results)
-        testing.save_image(input_image, Path("model_test.png"))
-        testing.save_json(results, Path("model_test.json"))
+        utils.draw_results(input_image, results)
+        utils.save_image(input_image, Path("model_test.png"))
+        utils.save_json(results, Path("model_test.json"))
 
     return results, next_id
