+++ sketchlogic/model/utils.py
